[APPS-5] Fixed SprykMainForm for Spryks with arguments, fixed process run for docker environment.
